docker-mcp-reports/requirements.txt
# Основные зависимости
requests>=2.31.0

# MCP (Model Context Protocol)
mcp>=1.0.0
openai>=1.0.0
anthropic>=0.34.0

# Логирование и утилиты
python-dotenv>=1.0.0

# Для работы с датами
python-dateutil>=2.8.0

# Для валидации данных
pydantic>=2.0.0

# Для работы с cron (планировщик задач)
python-crontab>=3.0.0

# Для работы с JSON
ujson>=5.8.0

# Для работы с HTTP
httpx>=0.25.0
flask>=2.3.0
flask-cors>=4.0.0

# Для работы с асинхронностью
aiohttp>=3.9.0

# Для работы с файлами
pathlib2>=2.3.0

# Для работы с конфигурацией
pyyaml>=6.0.0

# Для работы с логированием
colorlog>=6.8.0

# Для работы с тестированием
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Для работы с типизацией
mypy>=1.7.0
types-requests>=2.31.0

# Для работы с форматированием кода
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
